+++ src/services.ts
@@ -12,6 +12,7 @@ import {DeviceControlService} from './services/ios/device-control/index.js';
 import {CoreDeviceInfoService} from './services/ios/device-info/index.js';
 import DiagnosticsService from './services/ios/diagnostic-service/index.js';
 import {DVTSecureSocketProxyService} from './services/ios/dvt/index.js';
+import {ActivityTraceTap} from './services/ios/dvt/instruments/activity-trace-tap.js';
 import {ApplicationListing} from './services/ios/dvt/instruments/application-listing.js';
 import {ConditionInducer} from './services/ios/dvt/instruments/condition-inducer.js';
 import {DeviceInfo} from './services/ios/dvt/instruments/device-info.js';
@@ -253,6 +254,7 @@ export async function startDVTService(udid: string): Promise<DVTInstruments> {
     processControl: new ProcessControl(dvtService),
     sysmontap: new Sysmontap(dvtService),
     energyMonitor: new EnergyMonitor(dvtService),
+    activityTraceTap: new ActivityTraceTap(dvtService),
   };
 }